+++ TeslaCamPlayer/src/TeslaCamPlayer.BlazorHosted/Shared/Models/AppConfig.cs
@@ -4,5 +4,8 @@ public class AppConfig
 {
     public bool EnableDelete { get; set; }
     public string SpeedUnit { get; set; }
+    public string TimeFormat { get; set; }
+    public string DateFormat { get; set; }
+    public int UiBlurPx { get; set; }
 }